Improve display of network information
