+++ src/App.tsx
@@ -140,6 +140,51 @@ function MainWindow() {
 
 const theme = extendTheme({
   colors: { black: '#12141c' },
+  components: {
+    Code: {
+      sizes: {
+        xs: {
+          fontSize: 'xs',
+        },
+        sm: {
+          fontSize: 'sm',
+        },
+        md: {
+          fontSize: 'md',
+        },
+        lg: {
+          fontSize: 'lg',
+        },
+        xl: {
+          fontSize: 'xl',
+        },
+        '2xl': {
+          fontSize: '2xl',
+        },
+        '3xl': {
+          fontSize: '3xl',
+        },
+        '4xl': {
+          fontSize: '4xl',
+        },
+        '5xl': {
+          fontSize: '5xl',
+        },
+        '6xl': {
+          fontSize: '6xl',
+        },
+        '7xl': {
+          fontSize: '7xl',
+        },
+        '8xl': {
+          fontSize: '8xl',
+        },
+        '9xl': {
+          fontSize: '9xl',
+        },
+      },
+    },
+  },
   styles: {
     global: (props: { colorMode: string; theme: string }) => ({
       'div.navbar': {
