+++ src/shared/layout.tsx
@@ -1,8 +1,13 @@
 import React, { ReactNode } from 'react';
 import { Box, Flex, useColorModeValue } from '@chakra-ui/react';
 
+// Containers
+
 interface ContainerProps {
   children: ReactNode;
+}
+interface OptionalContainerProps {
+  children?: ReactNode;
 }
 export function SubmenuContainer({ children }: ContainerProps): JSX.Element {
   return (
@@ -43,4 +48,77 @@ export function ContentContainer({
 }
 ContentContainer.defaultProps = {
   pad: false,
+};
+
+// Cards
+
+interface CardProps extends ContainerProps {
+  width?: string | number;
+}
+export function Card({ children, width }: CardProps): JSX.Element {
+  return (
+    <Box
+      background={useColorModeValue('gray.100', 'gray.900')}
+      p={{ base: 4, sm: 8 }}
+      rounded="xl"
+      overflow="hidden"
+      width={width}
+    >
+      {children}
+    </Box>
+  );
+}
+Card.defaultProps = {
+  width: undefined,
+};
+export function CardHeader({ children }: ContainerProps): JSX.Element {
+  return (
+    <Box
+      background={useColorModeValue('gray.300', 'gray.700')}
+      mt={{ base: -4, sm: -8 }}
+      mx={{ base: -4, sm: -8 }}
+      pt={4}
+      pb={2}
+      px={4}
+    >
+      {children}
+    </Box>
+  );
+}
+export function CardToolbar({ children }: ContainerProps): JSX.Element {
+  return (
+    <Box
+      background={useColorModeValue('gray.300', 'gray.700')}
+      mx={{ base: -4, sm: -8 }}
+      px={4}
+      pb={4}
+    >
+      {children}
+    </Box>
+  );
+}
+export function CardBody({ children }: ContainerProps): JSX.Element {
+  return (
+    <Box mx={{ base: -4, sm: -8 }} pt={4} pb={2} px={4}>
+      {children}
+    </Box>
+  );
+}
+export function CardFooter({ children }: OptionalContainerProps): JSX.Element {
+  if (children === undefined) {
+    return (
+      <Box mb={{ base: -4, sm: -8 }} mx={{ base: -4, sm: -8 }} p={2}>
+        {children}
+      </Box>
+    );
+  }
+
+  return (
+    <Box mb={{ base: -4, sm: -8 }} mx={{ base: -4, sm: -8 }} p={4}>
+      {children}
+    </Box>
+  );
+}
+CardFooter.defaultProps = {
+  children: undefined,
 };
