+++ src/zerotier/networks/dns.ts
@@ -1,0 +1,3 @@
+const DNS_ZT_NETWORK_KEY = 'zerotier-net-id';
+
+export default DNS_ZT_NETWORK_KEY;
